backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
pypdf
sentence-transformers
google-generativeai
